Remove dedicated bundler classes, always bundle complete package
By always bundling the complete package directory, we can remove the
undocumented "explicit" bundler and its related custom autoload path
detection. This also means PSR-4 and any other autoload standard is now
supported out-of-the-box.
