--- src/Clue/PharComposer/Package/Bundler/BundlerInterface.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Clue\PharComposer\Package\Bundler;
-
-/**
- * The Bundler is responsible for creating a Bundle containing all files which belong to a given package
- */
-interface BundlerInterface
-{
-   /**
-     * returns a bundle
-     *
-     * @return \Clue\PharComposer\Package\Bundle
-     */
-    public function bundle();
-}
